requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
openai
pydantic
python-docx
pypdf
streamlit
python-multipart
reportlab 
plotly